feat(pwa/ff): job add modal
